+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.0.8.2 (Upcoming)
 
 * Allow associativity in search-time geospatial transformations
+* Simplify naming of static methods for creating geospatial transformations
 
 ## 3.0.8.1 (August 02, 2016)